requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
certifi==2023.5.7
charset-normalizer==3.2.0
Django==4.0
djangorestframework==3.14.0
exceptiongroup==1.1.2
freezegun==1.2.2
gunicorn==20.1.0
idna==3.4
iniconfig==2.0.0
packaging==23.1
pluggy==1.2.0
psycopg2==2.9.6
pytest==7.4.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing-extensions==4.7.1
urllib3==2.0.3
